added card design for tablevote
