+++ frontend/src/components/appointmentsettings/Appoint.scss
@@ -78,6 +78,10 @@
 
 }
 
+.addElements {
+  margin:30px;
+}
+
 .menuIcons {
   width: 300px;
   display: flex;
