--- frontend/src/components/appointmentsettings/AppointThree/AppointThreeSearchRestaurant.tsx
@@ -1,103 +1,0 @@
-import * as React from 'react';
-import Button from '@mui/material/Button';
-import {styled} from '@mui/material/styles';
-import Dialog from '@mui/material/Dialog';
-import DialogTitle from '@mui/material/DialogTitle';
-import DialogContent from '@mui/material/DialogContent';
-import DialogActions from '@mui/material/DialogActions';
-import IconButton from '@mui/material/IconButton';
-import CloseIcon from '@mui/icons-material/Close';
-import Typography from '@mui/material/Typography';
-import SearchIcon from '@mui/icons-material/Search';
-
-const BootstrapDialog = styled(Dialog)(({theme}) => ({
-    '& .MuiDialogContent-root': {
-        padding: theme.spacing(2),
-    },
-    '& .MuiDialogActions-root': {
-        padding: theme.spacing(1),
-    },
-}));
-
-export interface DialogTitleProps {
-    id: string;
-    children?: React.ReactNode;
-    onClose: () => void;
-}
-
-const BootstrapDialogTitle = (props: DialogTitleProps) => {
-    const {children, onClose, ...other} = props;
-
-    return (
-        <DialogTitle sx={{m: 0, p: 2}} {...other}>
-            {children}
-            {onClose ? (
-                <IconButton
-                    aria-label="close"
-                    onClick={onClose}
-                    sx={{
-                        position: 'absolute',
-                        right: 8,
-                        top: 8,
-                        color: (theme) => theme.palette.grey[500],
-                    }}
-                >
-                    <CloseIcon/>
-                </IconButton>
-            ) : null}
-        </DialogTitle>
-    );
-};
-
-export default function AppointThreeSearchRestaurant() {
-    const [open, setOpen] = React.useState(false);
-
-    const handleClickOpen = () => {
-        setOpen(true);
-    };
-
-    const handleClose = () => {
-        setOpen(false);
-    };
-
-    return (
-        <div>
-            <Button variant="outlined" onClick={handleClickOpen}> <SearchIcon/> Search </Button>
-
-            <BootstrapDialog
-                onClose={handleClose}
-                aria-labelledby="customized-dialog-title"
-                open={open}
-            >
-
-                <BootstrapDialogTitle id="customized-dialog-title" onClose={handleClose}>
-                    Modal title
-                </BootstrapDialogTitle>
-
-                <DialogContent dividers>
-                    <Typography gutterBottom>
-                        Cras mattis consectetur purus sit amet fermentum. Cras justo odio,
-                        dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac
-                        consectetur ac, vestibulum at eros.
-                    </Typography>
-                    <Typography gutterBottom>
-                        Praesent commodo cursus magna, vel scelerisque nisl consectetur et.
-                        Vivamus sagittis lacus vel augue laoreet rutrum faucibus dolor auctor.
-                    </Typography>
-                    <Typography gutterBottom>
-                        Aenean lacinia bibendum nulla sed consectetur. Praesent commodo cursus
-                        magna, vel scelerisque nisl consectetur et. Donec sed odio dui. Donec
-                        ullamcorper nulla non metus auctor fringilla.
-                    </Typography>
-                </DialogContent>
-
-                <DialogActions>
-                    <Button autoFocus onClick={handleClose}>
-                        Add Restaurant
-                    </Button>
-                </DialogActions>
-            </BootstrapDialog>
-
-        </div>
-    )
-}
