+++ frontend/src/components/appointmentsettings/AppointThree/Subcomponents/DialogTitle.tsx
@@ -1,0 +1,45 @@
+import {styled} from "@mui/material/styles";
+import Dialog from "@mui/material/Dialog";
+import * as React from "react";
+import DialogTitle from "@mui/material/DialogTitle";
+import IconButton from "@mui/material/IconButton";
+import CloseIcon from "@mui/icons-material/Close";
+
+export const BootstrapDialog = styled(Dialog)(({theme}) => ({
+    '& .MuiDialogContent-root': {
+        padding: theme.spacing(2),
+    },
+    '& .MuiDialogActions-root': {
+        padding: theme.spacing(1),
+    },
+}));
+
+export interface DialogTitleProps {
+    id: string;
+    children?: React.ReactNode;
+    onClose: () => void;
+}
+
+export const BootstrapDialogTitle = (props: DialogTitleProps) => {
+    const {children, onClose, ...other} = props;
+
+    return (
+        <DialogTitle sx={{m: 0, p: 2}} {...other}>
+            {children}
+            {onClose ? (
+                <IconButton
+                    aria-label="close"
+                    onClick={onClose}
+                    sx={{
+                        position: 'absolute',
+                        right: 8,
+                        top: 8,
+                        color: (theme) => theme.palette.grey[500],
+                    }}
+                >
+                    <CloseIcon/>
+                </IconButton>
+            ) : null}
+        </DialogTitle>
+    );
+};
