+++ frontend/src/components/general/DisplayMenus.scss
@@ -42,7 +42,7 @@
 
 .sidebarText {
   display: flex;
-  justify-content: start;
+  justify-content: flex-start;
   align-items: center;
   padding: 8px 0px 8px 16px;
   list-style: none;
@@ -74,7 +74,7 @@
     width: 100%;
     height: 80px;
     display: flex;
-    justify-content: start;
+    justify-content: flex-start;
     align-items: center;
 
     .icons-menu-close {
